+++ packages/next-common/components/pages/components/gov2/sidebar/status/preparation/depositPopup.js
@@ -5,12 +5,18 @@ import { toPrecision } from "next-common/utils";
 import React, { useCallback } from "react";
 import PopupLabel from "next-common/components/popup/label";
 import Input from "next-common/lib/input";
-import SignerWithBalance from "next-common/components/signerPopup/signerWithBalance";
+import Signer from "next-common/components/popup/fields/signerField";
 import TxSubmissionButton from "next-common/components/common/tx/txSubmissionButton";
 import { usePopupParams } from "next-common/components/popupWithSigner/context";
 import { useContextApi } from "next-common/context/api";
 import { useReferendaPallet } from "next-common/context/referenda/pallet";
 import CurrencyInput from "next-common/components/currencyInput";
+import BalanceProvider, {
+  useBalanceContext,
+} from "next-common/context/balance";
+import { GreyPanel } from "next-common/components/styled/containers/greyPanel";
+import { colorStyle, PromptTypes } from "next-common/components/scrollPrompt";
+import BigNumber from "bignumber.js";
 
 function PopupContent() {
   const { onClose } = usePopupParams();
@@ -27,7 +33,7 @@ function PopupContent() {
 
   return (
     <>
-      <SignerWithBalance />
+      <SignerTransferableBalance />
       <div>
         <PopupLabel text={"Referendum ID"} />
         <Input value={referendumIndex} disabled={true} />
@@ -40,7 +46,7 @@ function PopupContent() {
           symbol={node?.symbol}
         />
       </div>
-
+      <BalanceTip />
       <TxSubmissionButton getTxFunc={getTxFunc} onClose={onClose} />
     </>
   );
@@ -49,7 +55,46 @@ function PopupContent() {
 export default function DepositPopup(props) {
   return (
     <PopupWithSigner title="Place decision deposit" {...props}>
-      <PopupContent />
+      <BalanceProvider>
+        <PopupContent />
+      </BalanceProvider>
     </PopupWithSigner>
+  );
+}
+
+function SignerTransferableBalance() {
+  const { transferrable, loading } = useBalanceContext() || {};
+  return (
+    <Signer
+      showTransferable
+      balance={transferrable}
+      isBalanceLoading={loading}
+    />
+  );
+}
+
+function BalanceTip() {
+  const { transferrable, loading } = useBalanceContext() || {};
+  const { trackInfo: track } = usePostOnChainData();
+
+  if (loading || !track) {
+    return null;
+  }
+
+  const isEnough = BigNumber(transferrable).gte(track.decisionDeposit);
+
+  return (
+    <GreyPanel
+      className="text14Medium px-4 py-[10px]"
+      style={
+        isEnough
+          ? colorStyle[PromptTypes.SUCCESS]
+          : colorStyle[PromptTypes.WARNING]
+      }
+    >
+      {isEnough
+        ? "Your balance is enough to pay the deposit."
+        : "Your balance is not enough to pay the deposit."}
+    </GreyPanel>
   );
 }