+++ packages/next-common/components/profile/bio/index.jsx
@@ -21,6 +21,7 @@ import RightPanelContainer from "next-common/components/profile/bio/rightPanelCo
 import { useIsMobile } from "next-common/components/overview/accountInfo/components/accountBalances";
 import UserAccountProvider from "next-common/context/user/account";
 import AccountInfoPanel from "next-common/components/profile/bio/accountInfoPanel";
+import BioContainer from "./bioContainer";
 
 const Username = styled.span`
   font-weight: 700;
@@ -71,28 +72,54 @@ export const DisplayUser = ({ id, className = "" }) => {
   return <Username>{id}</Username>;
 };
 
-export const DisplayUserAddress = ({
+export const CopyableAddress = ({ address, ellipsisAddress = false }) => {
+  if (!address) {
+    return null;
+  }
+  const maybeEvmAddress = tryConvertToEvmAddress(address);
+
+  const displayAddress = ellipsisAddress
+    ? addressEllipsis(maybeEvmAddress)
+    : maybeEvmAddress;
+  return (
+    <Copyable copyText={maybeEvmAddress}>
+      <Tertiary>{displayAddress}</Tertiary>
+    </Copyable>
+  );
+};
+
+export const AccountAdditional = ({
   address,
-  className = "",
   showLinks = true,
-  ellipsisAddress = false,
-  extra = null,
-  accountLinksClassName = "",
+  children = null,
+  className = "",
 }) => {
   if (!address) {
     return null;
   }
   const maybeEvmAddress = tryConvertToEvmAddress(address);
-  const displayAddress = ellipsisAddress ? addressEllipsis(address) : address;
+  return (
+    <div className={cn("inline-flex items-center", className)}>
+      {showLinks && <AccountLinks address={maybeEvmAddress} />}
+      {children}
+    </div>
+  );
+};
+
+export const DisplayUserAddress = ({
+  address,
+  className = "",
+  ellipsisAddress = false,
+  children = null,
+}) => {
+  if (!address) {
+    return null;
+  }
   return (
     <AddressWrapper className={className}>
-      <Copyable copyText={maybeEvmAddress}>
-        <Tertiary>{displayAddress}</Tertiary>
-      </Copyable>
-      <div className={cn("inline-flex items-center", accountLinksClassName)}>
-        {showLinks && <AccountLinks address={maybeEvmAddress} />}
-        {extra}
-      </div>
+      <CopyableAddress address={address} ellipsisAddress={ellipsisAddress} />
+      <BioContainer />
+      {children}
     </AddressWrapper>
   );
 };