+++ packages/next/pages/_error.jsx
@@ -59,6 +59,7 @@ ErrorPage.getInitialProps = async ({ req, res, err }) => {
     address: user?.address,
     code: statusCode,
     source: isServerError ? "server" : "client",
+    userAgent: isServerError ? "" : window?.navigator?.userAgent,
     error: err.message,
     stack: err.stack,
   };